Tester med ny featuretoggle-metode
