+++ mock-data/unleash.json
@@ -1,6 +1,4 @@
 {
-  "dittnav.nytt-dinesakerpanel-testpool": true,
-  "dittnav.nytt-dinesakerpanel-ab": true,
   "dittnav.hendelser": false,
   "dittnav.ny-backend": true
 }
