+++ src/js/Api.js
@@ -4,18 +4,6 @@ const redirectToLogin = () => {
   window.location.assign(`${Config.dittNav.LOGINSERVICE}&redirect=${window.location.href}`);
 };
 
-const fetchUnleashFeatures = (features) => {
-  const fString = features.map(f => `feature=${f}`);
-  const URL = `${Config.dittNav.CONTEXT_PATH}/api/feature`;
-  return Promise.race([
-    fetch(`${URL}?${fString.join('&')}`, { method: 'GET' })
-      .then(r => r.json()),
-    new Promise((_, reject) => setTimeout(() => {
-      const message = `Couldnt wait for unleash longer than ${Config.UNLEASH_TIMEOUT} msec`;
-      return reject(new Error(message));
-    }, Config.UNLEASH_TIMEOUT))]);
-};
-
 const fetchJSON = (url) => new Promise((res, rej) => {
   fetch(url, { method: 'GET', credentials: 'include' })
     .then(r => {
@@ -73,7 +61,6 @@ const fetchSakstema = () => fetchJSON(Config.dittNav.DITTNAV_SAKSTEMA_URL);
 const fetchHendelser = () => fetchJSON(`${Config.dittNav.DITTNAV_HENDELSER_URL}`);
 
 export default {
-  fetchUnleashFeatures,
   checkAuth,
   checkApiStatus,
   fetchOppfolging,
