+++ src/js/NavApp.js
@@ -5,7 +5,7 @@ import { IntlProvider, addLocaleData } from 'react-intl';
 import nb from 'react-intl/locale-data/nb';
 import en from 'react-intl/locale-data/en';
 import nn from 'react-intl/locale-data/nn';
-import FeatureTogglesProvider from './components/FeatureTogglesProvider';
+import FeatureTogglesProvider from './components/FeatureToggles';
 
 addLocaleData([...nb, ...en, ...nn]);
